fix: move database initialization into scripts dir
