+++ src/database/index.ts
@@ -1,5 +1,5 @@
 import { DatabaseConnection } from './connection';
-import { initializeDatabase } from './initialize-db';
+import { initializeDatabase } from '../../scripts/initialize-db';
 import { repositories } from './repositories';
 
 export {
